move find_max_cell to max_cell submodule; add test
